Refactor file extension, remove hardcoded metrics display
The file extension for the joblib dump has been changed from '.jolib' to '.joblib' for standardization. Additionally, the hardcoded metrics display has been removed from the notebook. This is a necessity since the metrics will be different in different runs and hence it is not productive to hard code them in the notebook. The start and end execution times for code blocks have also been adjusted.